renaming and cleanup
