+++ main.go
@@ -3,18 +3,17 @@ package main
 import (
 	"context"
 
+	"github.com/viam-soleng/mqtt-client/mqttclient"
 	"go.viam.com/rdk/components/sensor"
 	"go.viam.com/rdk/logging"
 	"go.viam.com/rdk/module"
 	"go.viam.com/utils"
-
-	myMQTTClient "github.com/williamjhyland/MQTT_Integration_Example/MQTT"
 )
 
 func main() {
 	// NewLoggerFromArgs will create a logging.Logger at "DebugLevel" if
 	// "--log-level=debug" is an argument in os.Args and at "InfoLevel" otherwise.
-	utils.ContextualMain(mainWithArgs, module.NewLoggerFromArgs("My Go MQTT Module"))
+	utils.ContextualMain(mainWithArgs, module.NewLoggerFromArgs("Viam Solutions Engineering MQTT Client Module"))
 }
 
 func mainWithArgs(ctx context.Context, args []string, logger logging.Logger) (err error) {
@@ -26,7 +25,7 @@ func mainWithArgs(ctx context.Context, args []string, logger logging.Logger) (er
 
 	// Models and APIs add helpers to the registry during their init().
 	// They can then be added to the module here.
-	err = myMod.AddModelFromRegistry(ctx, sensor.API, myMQTTClient.Model)
+	err = myMod.AddModelFromRegistry(ctx, sensor.API, mqttclient.Model)
 	if err != nil {
 		return err
 	}
